a few minor changes to service provider
